Clean up of app.js; moved apiKey and firebase config to private file in gitignore
