+++ js/info.js
@@ -1,0 +1,12 @@
+const firebaseConfig = {
+    apiKey: "AIzaSyCLUor0R1QCzEdIo0FEqH3VUOXWoitujDQ",
+    authDomain: "test-19cf1.firebaseapp.com",
+    projectId: "test-19cf1",
+    storageBucket: "test-19cf1.appspot.com",
+    messagingSenderId: "407722304456",
+    appId: "1:407722304456:web:1dfe1d12f2b60e9a3816fd"
+};
+
+const apiKey = 'aIIJ5p83TsAOWEXASdgbJYiNjZ1kSNW0';
+
+export { firebaseConfig, apiKey }
